feat(idp): implement initial IDP server
