+++ Cargo.toml
@@ -60,3 +60,5 @@ reqwest = "0.12.8"
 base64 = "0.22"
 sqlx = "0.8"
 bon = "2.3"
+directories = "5.0"
+validator = { version = "0.19", features = ["derive"] }
