--- crates/tessera-authentication/default_config.toml
@@ -1,1 +1,0 @@
-port = 8080
