+++ crates/tessera-authority/Cargo.toml
@@ -14,8 +14,8 @@ tracing = { workspace = true }
 tracing-subscriber = { workspace = true }
 clap = { version = "4.5.19", features = ["cargo", "derive"] }
 axum = "0.7.7"
+directories = { workspace = true }
 config = "0.14.0"
-directories = "5.0"
 serde = { workspace = true }
 tessera-abe = { workspace = true }
 tessera-storage = { workspace = true }
