+++ crates/tessera-idp/src/application.rs
@@ -1,0 +1,7 @@
+pub struct Application;
+
+impl Application {
+    pub fn new() -> Self {
+        Self
+    }
+}
